feat : GetNextName вынесен в DataBasePage
