+++ GUI/DataBases/FunctionDataBasePage.cs
@@ -214,7 +214,7 @@ namespace BazisGUI.DataBases
                 var dbPath = string.Empty;
                 var name = string.Empty;
                 dbPath = Directory.GetFiles(Application.StartupPath, "functions_draft.txt", SearchOption.AllDirectories)[0];
-                name = GetNextName("Новая_функция_");
+                name = GetNextName("Новая_функция_", Functions.Keys);
 
                 var dataSet = Loader.LoadDataBase(dbPath);
                 var function = ConvertToFunctions(dataSet);
@@ -319,29 +319,6 @@ namespace BazisGUI.DataBases
             base.OpenFileDB_Click(sender, e);
         }
 
-        private string GetNextName(string basePrefix)
-        {
-            var numbers = Functions.Keys
-                .Where(k => k.StartsWith(basePrefix))
-                .Select(k =>
-                {
-                    var suffix = k.Substring(basePrefix.Length);
-                    return int.TryParse(suffix, out int n) ? n : (int?)null;
-                })
-                .Where(n => n.HasValue)
-                .Select(n => n.Value)
-                .OrderBy(n => n)
-                .ToList();
-            int next = 1;
-            foreach (var n in numbers)
-            {
-                if (n == next)
-                    next++;
-                else
-                    break;
-            }
-            return $"{basePrefix}{next}";
-        }
         private FunctionDBData ConvertToFunctions(DataSet dataSet)
         {
             var functions = new FunctionDBData();
